Change around the end-to-end and deployment tests. Use an end-to-end test during the test job, and a much shorter end-to-end (to check all components work together) for the deployment test
