+++ .circleci/config.yml
@@ -67,7 +67,7 @@ jobs:
               cat tests/deployment/requirements.txt
               pip install -r tests/deployment/requirements.txt
               ./dist/grakn-core-all-linux/grakn console -k diagnosis -f kglib/utils/grakn/synthetic/examples/diagnosis/schema.gql
-              python -m unittest kglib.kgcn.examples.diagnosis.diagnosis_end_to_end_test
+              python -m unittest kglib.kgcn.examples.diagnosis.diagnosis_deployment_test
 
   release-approval:
     machine: true
