+++ kglib/kgcn/examples/diagnosis/diagnosis.py
@@ -51,21 +51,23 @@ def diagnosis_example(num_graphs=60,
         print(f'Found node types: {node_types}')
         print(f'Found edge types: {edge_types}')
 
-    ge_graphs = pipeline(graphs,
-                         tr_ge_split,
-                         node_types,
-                         edge_types,
-                         num_processing_steps_tr=num_processing_steps_tr,
-                         num_processing_steps_ge=num_processing_steps_ge,
-                         num_training_iterations=num_training_iterations,
-                         categorical_attributes=CATEGORICAL_ATTRIBUTES,
-                         )
+    ge_graphs, solveds_tr, solveds_ge = pipeline(graphs,
+                                                 tr_ge_split,
+                                                 node_types,
+                                                 edge_types,
+                                                 num_processing_steps_tr=num_processing_steps_tr,
+                                                 num_processing_steps_ge=num_processing_steps_ge,
+                                                 num_training_iterations=num_training_iterations,
+                                                 categorical_attributes=CATEGORICAL_ATTRIBUTES,
+                                                 )
 
     with session.transaction().write() as tx:
         write_predictions_to_grakn(ge_graphs, tx)
 
     session.close()
     client.close()
+
+    return solveds_tr, solveds_ge
 
 
 CATEGORICAL_ATTRIBUTES = {'name': ['meningitis', 'flu', 'fever', 'light-sensitivity']}
