+++ kglib/kgcn/examples/diagnosis/diagnosis_deployment_test.py
@@ -20,20 +20,9 @@
 import unittest
 
 from kglib.kgcn.examples.diagnosis.diagnosis import diagnosis_example
-from kglib.utils.grakn.test.base import GraknServer
-
-TEST_KEYSPACE = "diagnosis"
-TEST_URI = "localhost:48555"
 
 
 class TestDiagnosisExample(unittest.TestCase):
-    def setUp(self):
-        self._gs = GraknServer()
-        self._gs.start()
-        self._gs.load_graql_file(TEST_KEYSPACE, '/kglib/kglib/utils/grakn/synthetic/examples/diagnosis/schema.gql')
-
-    def tearDown(self):
-        self._gs.stop()
 
     def test_example_runs_without_exception(self):
         diagnosis_example(num_graphs=6,
