+++ kglib/kgcn/examples/diagnosis/diagnosis_end_to_end_test.py
@@ -20,12 +20,25 @@
 import unittest
 
 from kglib.kgcn.examples.diagnosis.diagnosis import diagnosis_example
+from kglib.utils.grakn.test.base import GraknServer
+
+TEST_KEYSPACE = "diagnosis"
+TEST_URI = "localhost:48555"
 
 
 class TestDiagnosisExample(unittest.TestCase):
+    def setUp(self):
+        self._gs = GraknServer()
+        self._gs.start()
+        self._gs.load_graql_file(TEST_KEYSPACE, '/kglib/kglib/utils/grakn/synthetic/examples/diagnosis/schema.gql')
 
-    def test_example_runs_without_exception(self):
-        diagnosis_example()
+    def tearDown(self):
+        self._gs.stop()
+
+    def test_learning_is_done(self):
+        solveds_tr, solveds_ge = diagnosis_example()
+        self.assertGreaterEqual(solveds_tr[-1], 0.1)
+        self.assertGreaterEqual(solveds_tr[-1], 0.1)
 
 
 if __name__ == "__main__":
